add jwt extended and correct file path

+++ resources/item.py
@@ -1,5 +1,11 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required, current_identity
+from flask_jwt_extended import (
+    fresh_jwt_required,
+    jwt_required,
+    jwt_optional,
+    get_jwt_claims,
+    get_jwt_identity
+)
 
 from models.item import ItemModel
 
@@ -22,7 +28,7 @@ class Item(Resource):
     )
     
     
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -30,20 +36,23 @@ class Item(Resource):
         return {'message': 'Item not found'}, 404
 
 
-    @jwt_required()
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': 'An item with the name "{}" already exists.'.format(name)}, 400
         
         request_data = Item.parser.parse_args()
-        item = ItemModel(name, request_data['price'], request_data['store_id'])
+        item = ItemModel(name, **request_data)
         item.save_to_db()
 
         return item.json(), 201
     
     
-    @jwt_required()
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -51,14 +60,14 @@ class Item(Resource):
         return {'message': 'Item deleted'}, 200
     
     
-    @jwt_required()
+    @jwt_required
     def put(self, name):
         data = Item.parser.parse_args()
         
         item = ItemModel.find_by_name(name)
 
         if item is None:
-            item = ItemModel(name, data['price'], data['store_id'])
+            item = ItemModel(name, **data)
         else:
             item.price = data['price']
         
@@ -67,21 +76,15 @@ class Item(Resource):
         return item.json()
 
 class ItemList(Resource):
-    @jwt_required()
+    @jwt_optional
     def get(self):
-        """
-        print(current_identity.username)
-        connection = sqlite3.connect('data.db')
-        cursor = connection.cursor()
-        
-        query = 'SELECT * FROM items'
-        result = cursor.execute(query).fetchall()
-        
-        items = [{ 'name': row[0], 'price': row[1]} for row in result]
-        
-        connection.close()
-        
-        return items
-        """
-        return {'items': [item.json() for item in ItemModel.query.all()]}
+        # gives whatever we saved in the access token as the identity
+        # access_token = create_access_token(identity=user.id, fresh=True)
+        user_id = get_jwt_identity()
+        if user_id:
+            return {'items': [item.json() for item in ItemModel.query.all()]}, 200
+        return {
+            'items': [item.json()['name'] for item in ItemModel.query.all()],
+            'message': 'Login to view more info.'
+        }, 200
